refactor(assert): split typed array equality helper
